consumers/requirements.txt
celery
redis
ultralytics
onnxruntime
numpy
opencv-python-headless
python-dotenv
